LayoutTests/platform/ios-simulator/fast/canvas/canvas-crash-expected.txt

CONSOLE MESSAGE: Total canvas memory use exceeds the maximum limit (224 MB).
CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'ctx.putImageData')
